Run migration when app is ready
