+++ main/api/contacts/utils/migrateNodeAppBetaContacts.ts
@@ -3,6 +3,9 @@ import os from "os";
 import path from "path";
 
 import NeDBStorage from "nedb";
+import { v4 as uuidv4 } from "uuid";
+
+import { store } from "..";
 
 type BetaContact = {
   name: string;
@@ -11,7 +14,7 @@ type BetaContact = {
 
 let nedb: NeDBStorage;
 
-export function getNodeAppBetaContacts(): Promise<Array<BetaContact>> {
+function getNodeAppBetaContacts(): Promise<Array<BetaContact>> {
   const betaContactsDbPath = path.join(
     os.homedir(),
     "/.ironfish/node-app",
@@ -41,7 +44,7 @@ export function getNodeAppBetaContacts(): Promise<Array<BetaContact>> {
   });
 }
 
-export function deleteNodeAppBetaContacts(): void {
+function deleteNodeAppBetaContacts(): void {
   const betaContactsDbPath = path.join(
     os.homedir(),
     "/.ironfish/node-app",
@@ -51,4 +54,27 @@ export function deleteNodeAppBetaContacts(): void {
   if (fs.existsSync(betaContactsDbPath)) {
     fs.rmSync(betaContactsDbPath);
   }
+}
+
+export async function migrateNodeAppBetaContacts() {
+  const betaContacts = await getNodeAppBetaContacts();
+  const contacts = store.get("contacts", []);
+  const addressLookup = new Set(contacts.map((contact) => contact.address));
+
+  let order = 0;
+
+  for (const contact of betaContacts) {
+    if (!addressLookup.has(contact.address)) {
+      contacts.push({
+        id: uuidv4(),
+        name: contact.name,
+        address: contact.address,
+        order: order++,
+      });
+    }
+  }
+
+  store.set("contacts", contacts);
+  deleteNodeAppBetaContacts();
+  return contacts;
 }
